+++ packages/web3-eth-contract/CHANGELOG.md
@@ -387,3 +387,9 @@ Documentation:
 -   `defaultReturnFormat` was added to all methods that have `ReturnType` param. (#6947)
 
 ## [Unreleased]
+
+### Added
+
+-   `populateTransaction` was added to contract methods (#7124)
+
+-   Contract has `setTransactionMiddleware` and `getTransactionMiddleware` for automatically passing to `sentTransaction` for `deploy` and `send` functions (#7138)